docs: add first-pass security scan report for production readiness

+++ App/Views/home/partials/_ad-deals.php
@@ -36,7 +36,7 @@
                         <span class="inline-flex items-center rounded-full bg-white/20 px-2.5 py-0.5 text-xs font-bold">LIMITED TIME</span>
                     </div>
                     <h2 class="text-2xl md:text-3xl font-extrabold tracking-tight">Save up to 40% on selected items</h2>
-                    <p class="mt-1 text-white/80 text-sm">Don't miss these deals — <?= count($onSale) ?> products on sale right now.</p>
+                    <p class="mt-1 text-white/80 text-sm">Don't miss these deals — <?= (int) count($onSale) ?> products on sale right now.</p>
                 </div>
             </div>
             <div class="flex gap-3 shrink-0">
